+++ pyproject.toml
@@ -6,15 +6,14 @@ authors = ["Christian Hettlage <hettlage@saao.ac.za>"]
 
 [tool.poetry.dependencies]
 python = "^3.8"
-starlette = "^0.14.1"
+starlette = "^0.13"
+ariadne = "^0.12.0"
 uvicorn = "^0.12.2"
 starlette-jwt = "^0.1.9"
-PyMySQL = "^0.10.1"
 pandas = "^1.1.4"
-ariadne = "^0.12.0"
-starlette = "^0.13.0"
-uvicorn = "^0.12.2"
 python-dateutil = "^2.8.1"
+aiomysql = "^0.0.20"
+boolean = "^1.1.0"
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.1.2"
@@ -26,7 +25,6 @@ isort = "^5.6.4"
 mypy = "^0.790"
 mkdocs = "^1.1.2"
 mkdocs-material = "^6.1.5"
-pytz = "^2020.4"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
